fixed an error in the customer service with addCustomerPreferredSpecies

+++ src/main/java/com/example/demo/services/CustomerService.java
@@ -1,9 +1,11 @@
 package com.example.demo.services;
 
+import com.example.demo.exception.BadRequestException;
 import com.example.demo.models.Customer;
 import com.example.demo.models.enums.Species;
 import com.example.demo.repositories.CustomerRepo;
 import com.example.demo.repositories.SpeciesTableRepo;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -30,9 +32,11 @@ public class CustomerService {
 
     }
 
-    public void addNewCustomer(Customer customer) {
+    public ResponseEntity<Customer> addNewCustomer(Customer customer) {
 
         customerRepo.save(customer);
+
+        return ResponseEntity.ok(customer);
 
     }
 
@@ -45,32 +49,43 @@ public class CustomerService {
 
     }
 
-    public void updateCustomer(Customer returnCustomer, Customer customerDetails){
+    public ResponseEntity<Customer> updateCustomer(Customer returnCustomer, Customer customerDetails){
 
-        if(customerDetails.getFirstName()!=null){
-            returnCustomer.setFirstName(customerDetails.getFirstName());
-        }
-
-        returnCustomer.setLastName(customerDetails.getLastName());
-        returnCustomer.setAge(customerDetails.getAge());
-        returnCustomer.setLocation(customerDetails.getLocation());
+        if(customerDetails.getFirstName()!=null) returnCustomer.setFirstName(customerDetails.getFirstName());
+        if(customerDetails.getLastName()!=null) returnCustomer.setLastName(customerDetails.getLastName());
+        if(customerDetails.getAge() != null)  returnCustomer.setAge(customerDetails.getAge());
+        if(customerDetails.getLocation() != null)  returnCustomer.setLocation(customerDetails.getLocation());
 
         customerRepo.save(returnCustomer);
 
-
-
+        return ResponseEntity.ok(returnCustomer);
 
     }
 
     public void deleteCustomerPreferences(Long id){
-        customerRepo.deleteCustomerPreferences(id);
+
+
+            customerRepo.deleteCustomerPreferences(id);
+
+
+
     }
 
-    public void deleteCustomer(Customer returnCustomer){
-        customerRepo.delete(returnCustomer);
+    public ResponseEntity<String> deleteCustomer(Long customer_id){
+
+        if(customerRepo.findCustomerByID(customer_id) != null){
+
+            Customer returnCustomer = customerRepo.findCustomerByID(customer_id);
+            customerRepo.delete(returnCustomer);
+            return ResponseEntity.ok("Customer with id " + customer_id + " deleted");
+
+        }else{
+            throw new BadRequestException("Invalid customer id");
+        }
+
     }
 
-    public void addCustomerPreferredSpecies(Long customer_id, Species species){
+    public ResponseEntity<Customer> addCustomerPreferredSpecies(Long customer_id, Species species){
 
         if(speciesTableRepo.findBySpeciesEquals(species.toString()) == null){
 
@@ -78,7 +93,16 @@ public class CustomerService {
 
         }
 
-        customerRepo.addCustomerPreferredSpecies(customer_id, speciesTableRepo.findSpeciesID(species.toString()));
+        if(customerRepo.findCustomerByID(customer_id) != null){
+            customerRepo.addCustomerPreferredSpecies(customer_id, speciesTableRepo.findSpeciesID(species.toString()));
+
+            return ResponseEntity.ok(customerRepo.findCustomerByID(customer_id));
+        }else{
+            throw new BadRequestException("Invalid customer_id");
+        }
+
+
+
 
     }
